Added AccessRestrictionFilter and simplified authentication.
Replaced servlet container basic authentication with a custom servlet
filter which performs the same function.  The advantage to this is
that the servlet container is now divorced from the webapp.

The login service (realm) also simplified a great deal and removes its
Jetty dependencies.

Additionally, the basic authorization pop-up will be displayed as
needed based on the repository's access restriction.  This was
necessary for view-restricted repositories with the RSS feature.  Its
also necessary for completely open repositories as before it would
prompt for credentials.

Improved feed syndication feature.