Centralize default branch/HEAD resolution (issue 14)
If an object id was not specified Gitblit used HEAD to perform the
operation.  This breaks under some conditions like working with a
repository that does not have any commits on master but does have
commits on a vcs-import branch.

The new approach is to centralize the resolution of unspecified object
ids to a common method which resolves HEAD first but uses the most
recently modified branch if HEAD points to nothing.

This commit also includes a non-functional method for creating an empty
branch. I couldn't figure out how to make JGit create an orphaned
branch.