Tighter Gravatar integration. New window/tab on most activity links.
Each Gravatar thumbnail is now a link to the Gitblit Gravatar profile
page.  This page displays some of the profile information and contains a
link to the full Gravatar profile.

Activity page links now launch a new window/tab instead of targeting
"self" since it can be expensive to requery the activity information.