Use reflection to determine source repository index
Unfortunately Lucene's MultiReader class does not expose a public method
for determining which index reader a document comes from.  There is a
private method which determines this information.

My new approach is to use reflection to call this private method and
determine the actual source index at runtime instead of storing the
repository name in the index. This gets me out of the "rename
repository" problem at the expense of a little runtime performance.