Revised TimeUtils for localization
TimeUtils needs to output localized strings like "5 days ago" and "6 months".  In order to do this it needs a translation resource.  Additionally, that resource can not be static because the single Gitblit server can handle multiple connections in different locales/languages.

TimeUtils has changed from a collection of static methods to some static methods and some instance methods.  A TimeUtils is instantiated with an optional resource bundle which contains the preferred translation.  If the resourec bundle is null or the requested translation key does not exist, an English  default will be used.

This change required adjusting the signatures of several key methods and that cascaded out to adjusting those methods calls in many, many  classes.