Implemented optional create-on-push
If this feature is enabled then permitted users can automatically create a
repository when pushing to one that does not yet exist.  Permitted users
are administrators and any user with the CREATE role.

If the pushing account is an administrator, the created repository may be
located in any folder/project space.  This reposiory will inherit the server's
default access restriction and authorization control.  The repository owner
will be the pushing account.

If the pushing account is a regular user with the CREATE role, the repository
can only be located in the account's personal folder (~username/myrepo.git).
This repository will be VIEW restricted and accessible by NAMED users.  The
repository owner will be the pushing account.