Support serving repositories over the SSH transport
Gitblit would greatly benefit from an integrated SSH server.  This would
complete the transport trifecta. The code in this commit was mostly
authored by Eric Myrhe but has been rebased and tweaked to follow the
current conventions in the master branch.  It is incomplete and not
ready for use but serves as a great starting point for implementing SSH
support.