Using Gitblit for Local Repos
I really liked Gitblit over other alternatives because of an easy installation process but I did not find suitable information about using it for local repos. So I decided to write a post for it. If you want then you can include this as I just want other users to help.