Add unit test for HtpasswdUserService.
Add unit test class for Apache htpasswd user service.
Add method getNumberUsers() to HtpasswdUserService which is only for use by
the unit test.
Change access to GitBlit.getFileOrFilder(String, String) to
GitBlit.getFileOrFolder(string) because the former would result in a
segementation fault during unit testing as the settings are not present in
GitBlit at that time.
Move setting of this.htpasswdFile from constructor to setup() for the same
reason as above.