Stop supporting crypt() and plain text passwords at the same time.
Disable support of plain text and crypt() encrypted passwords at the same
time. This caused problems because they both don't have an identifying
prefix in the htpasswd file. This means that when someone gleams the
crypt() hashes from the file she could log in using the hash as it would
successfully test against the stored hash using plain text comparison.

Now only either crypt() is supported or plain text but not both. Following
the Apache server, plain text is only supported on Windows and NetWare.
Otherwise crypt() is supported.

An undocumented setting key "realm.htpasswd.supportPlaintextPasswords" is
used for unit testing and could potentially be used to override the
OS-specific behaviour.