Make realm.htpasswd.userfile changeable without restart.
Move reading of realm.htpasswd.userfile key to read() from setup(). This
means it can be changed during runtime without requiring a restart, at the
cost of checking the setting on every read(), which currently is every
authentication.

Changing the setting will clear the current htpasswd user base. That means
that setting it to a non-existing file will also disallow logins from any
user in the old file. This is a conscious decision since we allow the case
of not having a htpasswd file at all and only reling on the backing user
service.