Retrieve LDAP groups with dereferencing aliases.
Currently the LDAP user service will not dereference aliases when searching for groups. This patch enables
dereferencing aliases for the group search. This is benefitial if groups are defined in the DIT in a common
place but only certain ones shall play a role in Gitblit. These can now be linked under a group that can be
provided as search base for groups, without having to recreate the existing groups under the search base.

In addition, the new doSearch() method implemented in this patch also limits the attributes returned for
the gruop search to the "cn" attribute, which is the only one used. That prevents returning all the members
of the result groups, which can be a lot.