Define manager interfaces and update all of Gitblit to use managers
These manager interfaces define how the GitBlit singleton will
eventually be split into smaller component managers.  The Wicket app and
all servlets have been updated to request the needed managers.  There
are _very_ few method signature changes - although there are a handful.

This is a surgical sharding of responsibility based on a proof of
concept refactor.  Instead of random references to  GittBlit.self()
there are now precise references to the manager interface required to
accomplish some task.  Some tasks may require references to multiple
managers.

The code is now littered with calls to GitBlit.getManager(class) and
those familiar with the code-base will no doubt notice the duplication
of methods from IUserService in IUserManager and the addition of
implementation methods in the GitBlit context class.  When the
GitBlit class is broken apart and the existing external authentication
user service classes are refactored to AuthenticationService classes,
this will again simplify and flatten.  But in order to safely and
cleanly modularize the stable code-base we will have to live with a
little duplication for a short while.

Change-Id: I7314ec8acaab2dcc6092785ed4434cc09fdbbe16