Tickets module with patchset contributions
A basic issue tracker styled as a hybrid of Github and Bitbucket issues.
You may attach commits to an existing ticket or you can push a single
commit to create a *proposal* ticket.  The title and body of the commit
message become the subject and description of the proposal ticket.

Ticket tracker
--------------

The ticket tracker stores tickets as an append-only JSON journal in an
orphan branch within each repository.  The journals are deserialized,
and a ticket is built by applying the journal entries.  Tickets are
indexed using Apache Lucene and all queries and searches are executed
against this Lucene index.

There is one trade-off to this persistence design: it is non-relational.

What does that mean?  Gitblit stores user accounts within it's simple
`users.conf` file.  Users are identified by their username.  The
username is also the attribution hard-coded into each journal entry
(patchset revision, field changes, comment, etc).  If the username
changes in `users.conf`, the ticket data will not reflect that because
it is non-relational.

Here are a few reasons/justifications for this design choice:

1. commit identifications (author, committer, tagger) are non-relational
2. allows meaningful and complete ticket data to be distributed with the
repository
3. your favorite text editor can still be your administration tool

This combination of technologies allows Gitblit to continue to operate
without introducing an external persistence technology such as an SQL
database or a NoSQL store.

Having said that, the backend to the ticket tracker is pluggable and
there is a mostly complete Redis backend as a proof-of-concept.  For
those inclined, it should be relatively straight-forward to develop
other backends (H2, MongoDB, etc) as long as the journal design is
preserved.

Pushing Commits
---------------

Each push to a ticket is identified as a patchset revision and flagged
with a revision ref.  A patchset revision may build on the previous
patchsets (traditional fast-forward push) OR a patchset revision may
rewrite history (rebase, squash, rebase+squash, or amend).  Patchset
authors should not be afraid to polish, revise, and rewrite their code
before merging into the proposed branch.

Gitblit will create n+1 refs for each ticket, where n is the patchset
revision count.  These refs are `refs/ticket/{id}`, which are typically
used for pushing and pulling, and`refs/changes/{x}/{id}/{rev}` which
mark each pushed patchset revision.

**Normal collaboration workflow**

Patchsets for existing tickets should be pushed to: `refs/tickets/{id}`
where id is the integer ticket id.

    git push origin HEAD:refs/tickets/123

**Proposing a new change**

You may also create a new change proposal ticket just by pushing a
**single commit** to `refs/for/{branch}` where branch is the proposed
integration branch OR `refs/for/default` which specifies the default
integration branch configured for the repository.

    git push origin HEAD:refs/for/default

**Rebasing, Squashing, Amending**

Gitblit makes rebasing, squashing, and amending patchsets easy.  Because
Gitblit tracks the current patchset revision with `refs/ticket/{id}`,
pushing to that commit may result in a failure if a new revision is a
non-fast-forward push.  Only user's with (rewind) RW+ permission to the
repository can push a non-fast-forward update to `refs/tickets/{id}`.

There is an alternate ref that can be used by one of the patchset
authors or repository users with write (RW) access which will allow them
to rewrite a patchset.

    git push origin HEAD:refs/for/{id}

Pushing changes to this ref allows the patchset authors to rebase,
squash, or amend the patchset commits without requiring client-side use
of the *--force* flag on push AND without requiring RW+ permission to
the repository.  Since each patchset revision is tracked with a revision
ref it is easy to recover from an unwanted non-fast-forward patchset
revision.