Rollback to Servlet 2.5
Combining Dagger and Servlet 3 works really well on stock Tomcat and
Jetty but it is a troublesome combination on JEE containers with their
own ideas on how to instantiate classes.  JBoss AS 7 has been
particularly nasty and it is just simpler to scaleback and stay with
Servlet 2.5 than it is to fight all permuations of containers.

Instead of using constructor DI, the servlets and filters each have an
inject(ObjectGaph) method which is automatically called during
initialization.  Each servlet or filter is responsible for retrieving
the required dependency from the graph.  The Dagger object graph is
created in the context listener and stuffed into the context as an
attribute.

Change-Id: Ib5714584fe73e2a6b9c6fda12af080a43356cbda