Complete the simple review feature
A specific patchset revision may be reviewed using one of four scores:

1. approved (ready for merge)
2. looks good (changes look fine, but someone else must approve)
3. needs improvement (changes need refinement)
4. vetoed (changes are not allowed to be merged)

Any authenticated user who can clone the repository may review a
patchset and score with 2 or 3.  Both of these scores are non-binding.
An authenticated user with push permissions can score with 1 or 4.

By default you do not need to approve a patchset before it is merged,
however there is a per-repository setting to deactivate the merge button
if the current patchset does not have an approved review.  Similarly, if
a patchset revision is vetoed, no one will be able to use the merge
button in the web ui.

It should be noted that the approve/veto merge controls only apply to
the web ui.  An user with push permissions may circumvent a veto or a
missing approval by manually merging and pushing the merge to Gitblit.

Patchset updates.  If a patchset is updated or a rewritten patchset is
pushed, these reviews no longer apply; reviews are for a specific
patchset revision.  The reviews will still be part of the ticket history
and displayed on the activity tab.