Make displayname of users available in "access permission" lists.
Implemented a UserChoiceRenderer for the wicket palette, which uses the
display name from users. The lists are sorted with a new comparator
which compares on displayname.
Implemented tests for getDisplayName behaviour.

Implemented test for UserChoice behaviour. Refactored EditTeamPage to
use the UserChoice object. The UserChoice object is a lightweight object
with small serialization footprint. This avoids serialization of the
UserModel, which is a large object with a big serialization footprint.