Fire an event on received ticket branch updates to trigger reindex
It would be useful to be able to push a complete repository with a
refs/gitblit/tickets branch and have Gitblit index all those tickets.
Additionally, it would be nice for the mirroring service to recognize
an incoming update to this ref and fire an event that can be processed
by the BranchTicketService, if so configured, to reindex the tickets.