Add plugins/extension infrastructure
Plugins are stored in `${baseFolder}/plugins` and are loaded
during startup by the PluginManager.

A plugin defines it's metadata in META-INF/MANIFEST.MF:

    Plugin-Class: com.gitblit.plugins.cookbook.CookbookPlugin
    Plugin-Dependencies: foo, bar
    Plugin-Id: gitblit-plugin
    Plugin-Provider: John Doe
    Plugin-Version: 1.0

Plugins can define extension points that can be implemented
by other plugins and they can depend on other plugins:

    Plugin-Dependencies: foo, bar

During the load phase, a directed acyclic graph is built and the loading
order of the dependency chain is reversed using a topological sort;
parent followed by children. The parent plugin classloader is the
combined classloader of all parent plugins.

Change-Id: I738821fa2bff02a5dbe339a944cc7e3c4dd8e299