Merge overrides in addition to merging the normal settings
Some command-line parameters are ignored due to a bug in the merge
settings method which merges overrides into the normal settings map.
This works fine until the settings class decides to reload from the
pristine settings file which dumps the explicitly overrides.

The solution is to merge the overrides in addition to the standard
settings.