Italian translation.
When it comes to IT terminology, the Italian language borrows a lot of
terms from English, and it is often difficult to decide what is to be
translated and what is to be kept in English. I tried to follow the same
approach adopted by the translators of the pro-git book, so for example
"clone", "push" are not translated, while "branch" is translated (the
noun, not the git command).

I did my best to try to provide the best possible translation, but I had
not enough time to test all translations on screen.

Finally I noted that there are some duplicate keys e.g. gb.comment which
appears one with a capital C and once with a lowercase C.