provide spacing on the projects page listing
The projects page was showing the project listing directly under the navbar
with no spacing. This adds spacing to have the elements offset.

The table is showing using the common "repositories" class, however it can't
be changed directly since several other table displays use the same class, and
have adequate top margin spacing. We create a new class for the one-off use
on the projects page.